refactor: use u16

+++ crates/api/src/account/account_search_events.rs
@@ -72,7 +72,7 @@ pub struct AccountSearchEventsUseCase {
     pub sort: Option<CalendarEventSort>,
 
     /// Optional limit
-    pub limit: Option<u32>,
+    pub limit: Option<u16>,
 }
 
 #[derive(Debug)]
