+++ crates/api_structs/src/event/api.rs
@@ -334,9 +334,9 @@ pub mod search_events {
         #[ts(optional)]
         pub sort: Option<CalendarEventSort>,
 
-        /// Optional limit to use when searching events
+        /// Optional limit to use when searching events (u16)
         #[ts(optional)]
-        pub limit: Option<u32>,
+        pub limit: Option<u16>,
     }
 
     /// Part of the Request body for searching events for a user
